chore(hypridle): disable dpms
main monitor gets disabled in hyprland while the secondary one stays
enabled but turned off. Likely due to DP vs DVI connections.

this causes the main monitor re-enabled as second screen which messes up
with eww.

possible solution is to change connections on the GPU side so that the
main monitor is seen as a second one in hyprland.